DataCollection/requirements.txt
pyserial
coloredlogs
verboselogs
protobuf
numpy
h5py
